feat: wtf

+++ include/philo.h
@@ -6,7 +6,7 @@
 /*   By: ltreser <ltreser@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/20 16:42:34 by ltreser           #+#    #+#             */
-/*   Updated: 2024/11/12 01:53:30 by ltreser          ###   ########.fr       */
+/*   Updated: 2024/11/12 23:16:35 by ltreser          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,6 +97,6 @@ void    *monitor(void *arg);
 void	*routine(void *arg);
 void	serve_dinner(t_table *table);
 long	timestamp(t_philo *philo);
-int     check_t_last_meal(t_philo *philo);
+long     check_t_last_meal(t_philo *philo);
 
 #endif
